[core] layermanager folder
Move `LayerManager` and `LayerFactory` abstract classes to a dedicated folder.
