+++ src/mbgl/style/layer.cpp
@@ -110,63 +110,4 @@ const LayerTypeInfo* Layer::getTypeInfo() const noexcept {
 }
 
 } // namespace style
-
-optional<std::string> LayerFactory::getSource(const style::conversion::Convertible& value) const noexcept {
-    auto sourceValue = objectMember(value, "source");
-    if (!sourceValue) {
-        return nullopt;
-    }
-
-    optional<std::string> source = toString(*sourceValue);
-    if (!source) {
-        return nullopt;
-    }
-
-    return source;
-}
-
-bool LayerFactory::initSourceLayerAndFilter(style::Layer* layer, const style::conversion::Convertible& value) const noexcept {
-    auto sourceLayerValue = objectMember(value, "source-layer");
-    if (sourceLayerValue) {
-        optional<std::string> sourceLayer = toString(*sourceLayerValue);
-        if (!sourceLayer) {
-            return false;
-        }
-        layer->setSourceLayer(*sourceLayer);
-    }
-
-    auto filterValue = objectMember(value, "filter");
-    if (filterValue) {
-        style::conversion::Error error;
-        optional<style::Filter> filter = style::conversion::convert<style::Filter>(*filterValue, error);
-        if (!filter) {
-            return false;
-        }
-        layer->setFilter(*filter);
-    }
-
-    return true;
-}
-
-std::unique_ptr<style::Layer> LayerManager::createLayer(
-    const std::string& type, const std::string& id,
-    const style::conversion::Convertible& value, style::conversion::Error& error) noexcept {
-    if (LayerFactory* factory = getFactory(type)) {
-        auto layer = factory->createLayer(id, value);
-        if (!layer) {
-            error.message = "Error parsing a layer of type: " + type;
-        }
-        return layer;
-    }
-    error.message = "Unsupported layer type: " + type;
-    return nullptr;
-}
-
-std::unique_ptr<RenderLayer> LayerManager::createRenderLayer(Immutable<style::Layer::Impl> impl) noexcept {
-    LayerFactory* factory = getFactory(impl->getTypeInfo());
-    assert(factory);
-    return factory->createRenderLayer(std::move(impl));
-}
-
-
 } // namespace mbgl
